requirements.txt
python-dotenv
pyodbc
pdfplumber
pdf2image
pytesseract
Pillow
pyzbar
PyMuPDF
pytest
